refactor: files paths
