+++ lib/foundation/keyboard_controller.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:kiyibodi/keyboard_editing_value.dart';
+import 'package:kiyibodi/foundation/keyboard_editing_value.dart';
 
 import 'keyboard_input_type.dart';
