+++ lib/widgets/flexible_keyboard_input.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:kiyibodi/keyboard_input_type.dart';
+import 'package:kiyibodi/foundation/keyboard_input_type.dart';
 import 'package:kiyibodi/widgets/widgets.dart';
 
 class FlexibleKeyboardInput extends StatelessWidget {
