+++ lib/widgets/widgets.dart
@@ -1,3 +1,4 @@
 export 'keyboard_input.dart';
 export 'keyboard_input.dart';
 export 'keyboard_ink_well.dart';
+export 'keyboard.dart';
